+++ src/app/(pages)/userManagement/components/userPermissionsTable/UserPermissionsTable.test.tsx
@@ -24,7 +24,7 @@ jest.mock("@/app/backend/user-management", () => ({
   updateUserRole: jest.fn(),
 }));
 
-jest.mock("@/app/backend/seeding/service", () => ({
+jest.mock("@/app/backend/query-building/service", () => ({
   getConditionsData: jest.fn(),
 }));
