+++ src/app/assets/dibbs_db_seed_query.json
@@ -2,14 +2,12 @@
   "query": [
     {
       "author": "DIBBs",
-      "query_id": "cf580d8d-cc7b-4eae-8a0d-96c36f9222e3",
-      "date_created": "2024-12-10T15:45:30",
-      "date_last_modified": "2024-12-10T15:45:30",
-      "time_window_number": "30",
-      "time_window_unit": "days",
-      "query_name": "Cancer case investigation",
-      "conditions_list": "{2}",
-      "query_data": {
+      "queryId": "cf580d8d-cc7b-4eae-8a0d-96c36f9222e3",
+      "dateCreated": "2024-12-10T15:45:30",
+      "dateLastModified": "2024-12-10T15:45:30",
+      "queryName": "Cancer case investigation",
+      "conditionsList": "{2}",
+      "queryData": {
         "2": {
           "14_20240923": {
             "valueSetId": "14_20240923",
@@ -118,14 +116,13 @@
     },
     {
       "author": "DIBBs",
-      "query_id": "6edd14a2-ef78-4d8e-8509-0f87a7228d67",
-      "date_created": "2024-12-10T15:45:30",
-      "date_last_modified": "2024-12-10T15:45:30",
-      "time_window_number": "30",
-      "time_window_unit": "days",
-      "query_name": "Chlamydia case investigation",
-      "conditions_list": "{240589008}",
-      "query_data": {
+      "queryId": "6edd14a2-ef78-4d8e-8509-0f87a7228d67",
+      "dateCreated": "2024-12-10T15:45:30",
+      "dateLastModified": "2024-12-10T15:45:30",
+
+      "queryName": "Chlamydia case investigation",
+      "conditionsList": "{240589008}",
+      "queryData": {
         "240589008": {
           "2.16.840.1.113762.1.4.1146.999_20230602": {
             "valueSetId": "2.16.840.1.113762.1.4.1146.999_20230602",
@@ -2169,13 +2166,12 @@
     },
     {
       "author": "DIBBs",
-      "query_id": "facfefc1-dd39-4f84-9d91-e924e860ad1c",
-      "date_created": "2024-12-10T15:45:30",
-      "date_last_modified": "2024-12-10T15:45:30",
-      "time_window_number": "30",
-      "time_window_unit": "days",
-      "query_name": "Syphilis case investigation",
-      "query_data": {
+      "queryId": "facfefc1-dd39-4f84-9d91-e924e860ad1c",
+      "dateCreated": "2024-12-10T15:45:30",
+      "dateLastModified": "2024-12-10T15:45:30",
+
+      "queryName": "Syphilis case investigation",
+      "queryData": {
         "35742006": {
           "2.16.840.1.113762.1.4.1146.554_20191227": {
             "valueSetId": "2.16.840.1.113762.1.4.1146.554_20191227",
@@ -3350,17 +3346,16 @@
           }
         }
       },
-      "conditions_list": "{35742006}"
+      "conditionsList": "{35742006}"
     },
     {
       "author": "DIBBs",
-      "query_id": "73e1a777-49cb-4e19-bc71-8c3fd3ffda64",
-      "date_created": "2024-12-10T15:45:30",
-      "date_last_modified": "2024-12-10T15:45:30",
-      "time_window_number": "30",
-      "time_window_unit": "days",
-      "query_name": "Gonorrhea case investigation",
-      "query_data": {
+      "queryId": "73e1a777-49cb-4e19-bc71-8c3fd3ffda64",
+      "dateCreated": "2024-12-10T15:45:30",
+      "dateLastModified": "2024-12-10T15:45:30",
+
+      "queryName": "Gonorrhea case investigation",
+      "queryData": {
         "15628003": {
           "2.16.840.1.113762.1.4.1146.1036_20190605": {
             "valueSetId": "2.16.840.1.113762.1.4.1146.1036_20190605",
@@ -4894,17 +4889,16 @@
           }
         }
       },
-      "conditions_list": "{15628003}"
+      "conditionsList": "{15628003}"
     },
     {
       "author": "DIBBs",
-      "query_id": "c025a247-0129-4f0c-a2c6-7f3af08e06b4",
-      "date_created": "2024-12-10T15:45:30",
-      "date_last_modified": "2024-12-10T15:45:30",
-      "time_window_number": "30",
-      "time_window_unit": "days",
-      "query_name": "Newborn screening follow-up",
-      "query_data": {
+      "queryId": "c025a247-0129-4f0c-a2c6-7f3af08e06b4",
+      "dateCreated": "2024-12-10T15:45:30",
+      "dateLastModified": "2024-12-10T15:45:30",
+
+      "queryName": "Newborn screening follow-up",
+      "queryData": {
         "1": {
           "1_20240909": {
             "valueSetId": "1_20240909",
@@ -5929,7 +5923,7 @@
           }
         }
       },
-      "conditions_list": "{1}"
+      "conditionsList": "{1}"
     }
   ]
 }
